requirements.txt
opencv-python
pytesseract
pyttsx3
speechrecognition
google-generativeai
transformers
torch
ultralytics
python-dotenv
sounddevice
pyautogui
